# swagger.yaml
openapi: 3.0.3
info:
  title: Grymey
  description: >-
    # Grymey api documentation
  version: 1.0.0
  contact: { email: ammuftau74@gmail.com }
servers:
  - url: https://grymey.onrender.com/api/v1
    description: This is the original api URL
  - url: http://localhost:9000/api/v1
    description: This is my local Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /escrow:
    get:
      tags:
        - 🔄 escrow
      summary: Get escrow
      description: >-
        The API endpoint retrieves all escrow.
      operationId: getallescrow
      responses:
        '200':
          description: Get escrow
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '1126'
            Date:
              schema:
                type: string
                example: Thu, 01 Jun 2023 05:44:27 GMT
            ETag:
              schema:
                type: string
                example: W/"466-tNv1SNtB6cuxJH2yZW7lrDnlA34"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                  message:
                    type: string
                    example: Escrow fetched successfully
                  success:
                    type: boolean
                    example: true    
    post:
      tags:
        - 🔄 escrow
      summary: create escrow
      description: create escrow.
      operationId: createescrow
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                receiverId:
                  type: string
                  example: 689e5adc3e9ad666db548b6a
                description:
                  type: string
                  example: save some money for my school fees
                conditions:
                  type: string
                  example: save some money for my school fees
                expiresInDays:
                  type: string
                  example: '2025-09-30'
                amount:
                  type: string
                  example: 300000
      responses:
        '200':
          description: create escrow.
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '105'
            Date:
              schema:
                type: string
                example: Wed, 31 May 2023 19:56:15 GMT
            ETag:
              schema:
                type: string
                example: W/"69-7OD21C6OqsuAipwphfFVCzbFDQc"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties: {}
                  message:
                    type: string
                    example: escrow created successfully
                  success:
                    type: boolean
                    example: true
              examples:
                Forgot password request:
                  value:
                    data: {}
                    message: escrow created successfully
                    success: true
...
  /escrow/{escrowId}/release:
    post:
      tags:
        - 🔄 escrow
      summary: release escrow
      description: The API endpoint to release escrow request
      operationId: releaseescrow
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: bithday
                description:
                  type: string
                  example: general bithday
                amount:
                  type: string
                  example: 3000
            example:
              amount: 3000
              receiverId: 9399303iueiu983iu3
      responses:
        '200':
          description: Release escrow
          headers:
            Access-Control-Allow-Credentials:
              schema:
                type: string
                example: 'true'
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Connection:
              schema:
                type: string
                example: keep-alive
            Content-Length:
              schema:
                type: string
                example: '105'
            Date:
              schema:
                type: string
                example: Wed, 31 May 2023 19:56:15 GMT
            ETag:
              schema:
                type: string
                example: W/"69-7OD21C6OqsuAipwphfFVCzbFDQc"
            Keep-Alive:
              schema:
                type: string
                example: timeout=5
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties: {}
                  message:
                    type: string
                    example: Escrow release successfully
                  success:
                    type: boolean
                    example: true 
...